requirements.txt
pygame>=2.0.0
requests>=2.25.0
pyyaml>=6.0
psutil>=5.8.0
coverage>=6.0 